requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.3
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
pytz==2020.4
sqlparse==0.4.1
Werkzeug==1.0.1
